requirements.txt
Django==1.8.4
django-crispy-forms==1.5.2
django-debug-toolbar==1.3.2
django-extensions==1.5.7
django-haystack==2.4.0
django-model-utils==2.3.1
django-mptt==0.7.4
django-nvd3==0.8.2
django-tables2==1.0.4
django-bootstrap3==6.2.2
django-autocomplete-light==2.2.10
